Converting to markdown; let's do this shit!
I'm making my own build system, and it has an amazing name.  I
thought SCons would be overkill for simply converting some markdown
files to HTML, but that Bash would be just right, so here we are.
A build system is born, a README is updated, and a shitton of new
files are added.  That's life.