More ldt.sh hacking
This is turning into a more serious thing than I had originally
planned.  Though it uses a weird hack to print the command it's
about to run now, it seems to work very well and have no(?) adverse
side effects.  It is printing it to stderr, though, so that might
not be ideal.  I'll have to look into this further some time.